API Reference

API Reference

Index

Phase unwrapping

Background field correction

Dipole inversion

Binary masks

Kernels

Multi-echo

Other

Index

QSM.bet

QSM.crop_indices

QSM.crop_mask

QSM.dipole_kernel

QSM.erode_mask

QSM.erode_mask!

QSM.fastfftsize

QSM.fit_echo_linear

QSM.fit_echo_linear!

QSM.gradfp

QSM.gradfp!

QSM.gradfp_adj

QSM.gradfp_adj!

QSM.ismv

QSM.lap

QSM.lap!

QSM.laplace_kernel

QSM.lbv

QSM.nltv

QSM.padarray!

QSM.padfastfft

QSM.pdf

QSM.psf2otf

QSM.r2star_arlo

QSM.r2star_crsi

QSM.r2star_ll

QSM.r2star_numart2s

QSM.rts

QSM.sharp

QSM.smv_kernel

QSM.tikh

QSM.tkd

QSM.tsvd

QSM.tv

QSM.unpadarray

QSM.unpadarray!

QSM.unwrap_laplacian

QSM.vsharp

Phase unwrapping

QSM.unwrap_laplacian — Function

unwrap_laplacian(
    phas::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    solver::Symbol = :mgpcg
) -> typeof(similar(phas))

Laplacian phase unwrapping [1].

The Laplacian is computed using second order central finite differences on the complex phase. The resulting Poisson's equation can be solved with homogeneous Dirichlet (solver = :mgpcg), Neumann (:dct), or periodic (:fft) boundary conditions (BCs). Neumann and periodic BCs are imposed on the array while Dirichlet BCs are imposed on a region-of-interest (ROI) (mask). The boundary of the ROI is set such that values outside of it (mask = 0) are taken as boundary points and values inside of it (mask = 1) as interior points, ie. BC: uphas[!mask] = 0. This method combines phase unwrapping [1] and harmonic background field removing [2].

Arguments

phas::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: wrapped (multi-echo) phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

solver::Symbol = :mgpcg: solver for Poisson equation

:dct: homogeneous Neumann boundary condition on array

:fft: periodic boundary condition on array

:mgpcg: homogeneous Dirichlet boundary condition on mask (multigrid-preconditioned conjugate gradient method)

Returns

typeof(similar(phas)): unwrapped phase

References

[1] Schofield MA, Zhu Y. Fast phase unwrapping algorithm for interferometric applications. Optics letters. 2003 Jul 15;28(14):1194-6.

[2] Zhou D, Liu T, Spincemaille P, Wang Y. Background field removal by solving the Laplacian boundary value problem. NMR in Biomedicine. 2014 Mar;27(3):312-9.

Background field correction

QSM.ismv — Function

function ismv(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    r::Real = 2*maximum(vsz),
    tol::Real = 1e-3,
    maxit::Integer = 500,
    verbose::Bool = false,
) -> Tuple{typeof(similar(f)), typeof(similar(mask))}

Iterative spherical mean value method (iSMV) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size for smv kernel

Keywords

r::Real = 2*maximum(vsz): radius of smv kernel in units of vsz

tol::Real = 1e-3: stopping tolerance

maxit::Integer = 500: maximum number of iterations

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): background corrected local field/phase

typeof(similar(mask)): eroded binary mask

References

[1] Wen Y, Zhou D, Liu T, Spincemaille P, Wang Y. An iterative spherical mean value method for background field removal in MRI. Magnetic resonance in medicine. 2014 Oct;72(4):1065-71.

QSM.lbv — Function

lbv(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    atol::Real = sqrt(eps(Float64)),
    rtol::Real = sqrt(eps(Float64)),
    maxit::Integer = maximum(size(f)),
    verbose::Bool = false
) -> typeof(similar(f))

Laplacian boundary value problem (LBV) [1].

The Laplacian is computed using second order central finite differences. The resulting Poisson's equation is then solved inside an ROI (mask) with homogenous Dirichlet boundary condition (BC) using a multigrid-preconditioned conjugate gradient method. The boundary of the ROI is set such that values outside of it (mask = 0) are taken as boundary points and values inside of it (mask = 1) as interior points, ie. BC: fl[!mask] = 0.

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

atol::Real = sqrt(eps(Float64)): absolute stopping tolerance

rtol::Real = sqrt(eps(Float64)): relative stopping tolerance

maxit::Integer = maximum(size(f)): maximum number of cg iterations

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): background corrected local field/phase

References

[1] Zhou D, Liu T, Spincemaille P, Wang Y. Background field removal by solving the Laplacian boundary value problem. NMR in Biomedicine. 2014 Mar;27(3):312-9.

QSM.pdf — Function

pdf(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing,
    pad::NTuple{3, Integer} = (0, 0, 0),
    bdir::NTuple{3, Real} = (0, 0, 1),
    Dkernel::Symbol = :i,
    lambda::Real = 0,
    tol::Real = 1e-5,
    maxit::Integer = ceil(sqrt(numel(mask))),
    verbose::Bool = false
) -> typeof(similar(f))

Projection onto dipole fields (PDF) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size for dipole kernel

Keywords

W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing: data fidelity weights

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :i: dipole kernel method

lambda::Real = 0: regularization parameter

tol::Real = 1e-5: stopping tolerance for iterative solver

maxit::Integer = ceil(sqrt(length(mask))): maximum number of iterations for iterative solver

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): background corrected local field/phase

References

[1] Liu T, Khalidov I, de Rochefort L, Spincemaille P, Liu J, Tsiouris AJ, Wang Y. A novel background field removal method for MRI using projection onto dipole fields. NMR in Biomedicine. 2011 Nov;24(9):1129-36.

QSM.sharp — Function

sharp(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    r::Real = 18*minimum(vsz),
    thr::Real = 0.05,
) -> Tuple{typeof(similar(f)), typeof(similar(mask))}

Sophisticated harmonic artifact reduction for phase data (SHARP) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size for smv kernel

Keywords

r::Real = 18*minimum(vsz): radius of smv kernel in units of vsz

thr::Real = 0.05: threshold for high pass filter

Returns

typeof(similar(f)): background corrected local field/phase

typeof(similar(mask)): eroded binary mask

References

[1] Schweser F, Deistung A, Lehr BW, Reichenbach JR. Quantitative imaging of intrinsic magnetic tissue properties using MRI signal phase: an approach to in vivo brain iron metabolism?. Neuroimage. 2011 Feb 14;54(4):2789-807.

QSM.vsharp — Function

vsharp(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    r::AbstractVector{<:Real} = 18*minimum(vsz):-2*maximum(vsz):2*maximum(vsz),
    thr::Real = 0.05,
) -> Tuple{typeof(similar(f)), typeof(similar(mask))}

Variable kernels sophisticated harmonic artifact reduction for phase data (V-SHARP) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size for smv kernel

Keywords

r::AbstractVector{<:Real} = 18*minimum(vsz):-2*maximum(vsz):2*maximum(vsz): radii of smv kernels in mm

thr::Real = 0.05: threshold for high pass filter

Returns

typeof(similar(f)): background corrected local field/phase

typeof(similar(mask)): eroded binary mask

References

[1] Wu B, Li W, Guidon A, Liu C. Whole brain susceptibility mapping using compressed sensing. Magnetic resonance in medicine. 2012 Jan;67(1):137-47.

Dipole inversion

QSM.nltv — Function

nltv(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing,
    Wtv::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing,
    pad::NTuple{3, Integer} = (0, 0, 0),
    Dkernel::Symbol = :k,
    bdir::NTuple{3, Real} = (0, 0, 1),
    lambda::Real = 1e-3,
    rho::Real = 100*lambda,
    mu::Real = 1,
    tol::Real = 1e-3,
    maxit::Integer = 250,
    toln::Real = 1e-6,
    maxitn::Integer = 10,
    verbose::Bool = false,
) -> typeof(similar(f))

Nonlinear total variation deconvolution using ADMM [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing: data fidelity weights

Wtv::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, 5)}} = nothing: total variation weights

M = 3: same weights for all three gradient directions and all echoes

M = 4 = N: same weights for all three gradient directions, different weights for echoes

M = 5, (size(Wtv)[4,5] = [1 or N, 3]: different weights for each gradient direction

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

lambda::Real = 1e-3: regularization parameter

rho::Real = 100*lambda: Lagrange multiplier penalty parameter

mu::Real = 1: Lagrange multiplier penalty parameter (unused if W = nothing)

tol::Real = 1e-3: stopping tolerance

maxit::Integer = 250: maximum number of iterations

toln::Real = 1e-6: stopping tolerance for Newton method

maxitn::Integer = 10: maximum number of iterations for Newton method

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): susceptibility map

References

[1] Milovic C, Bilgic B, Zhao B, Acosta‐Cabronero J, Tejos C. Fast nonlinear susceptibility inversion with variational regularization. Magnetic resonance in medicine. 2018 Aug;80(2):814-21.

QSM.rts — Function

rts(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    pad::NTuple{3, Integer} = (0, 0, 0),
    Dkernel::Symbol = :k,
    bdir::NTuple{3, Real} = (0, 0, 1),
    lstol::Integer = 4,
    delta::Real = 0.15,
    mu::Real = 1e5,
    rho::Real = 10,
    tol::Real = 1e-2,
    maxit::Integer = 20,
    verbose::Bool = false,
) -> typeof(similar(f))

Rapid two-step dipole inversion with sparsity priors [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

lstol::Integer = 4: stopping tolerance (# of iterations) for lsmr solver

delta::Real = 0.15: threshold for ill-conditioned k-space region

mu::Real = 1e5: regularization parameter for tv minimization

rho::Real = 10: Lagrange multiplier penalty parameter

tol::Real = 1e-2: stopping tolerance

maxit::Integer = 20: maximum number of iterations

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): susceptibility map

References

[1] Kames C, Wiggermann V, Rauscher A. Rapid two-step dipole inversion for susceptibility mapping with sparsity priors. Neuroimage. 2018 Feb 15;167:276-83.

QSM.tikh — Function

tikh(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    pad::NTuple{3, Integer} = (0, 0, 0),
    bdir::NTuple{3, Real} = (0, 0, 1),
    Dkernel::Symbol = :k,
    lambda::Real = 1e-2,
    reg::Symbol = :gradient
) -> typeof(similar(f))

Tikhonov regularization [1].

\[ argmin_x ||Dx - f||_2^2 + \frac{λ}{2}||Γx||_2^2\]

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

lambda::Real = 1e-2: regularization parameter

reg::Symbol = :identity: regularization matrix Γ (:identity, :gradient, :laplacian)

Returns

typeof(similar(f)): susceptibility map

References

[1] Bilgic B, Chatnuntawech I, Fan AP, Setsompop K, Cauley SF, Wald LL, Adalsteinsson E. Fast image reconstruction with L2‐regularization. Journal of magnetic resonance imaging. 2014 Jul;40(1):181-91.

QSM.tkd — Function

tkd(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    pad::NTuple{3, Integer} = (0, 0, 0),
    bdir::NTuple{3, Real} = (0, 0, 1),
    Dkernel::Symbol = :k,
    thr::Real = 0.15,
) -> typeof(similar(f))

Truncated k-space division (TKD) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

thr::Real = 0.15: threshold for k-space filter

Returns

typeof(similar(f)): susceptibility map

References

[1] Shmueli K, de Zwart JA, van Gelderen P, Li TQ, Dodd SJ, Duyn JH. Magnetic susceptibility mapping of brain tissue in vivo using MRI phase data. Magnetic Resonance in Medicine: An Official Journal of the International Society for Magnetic Resonance in Medicine. 2009 Dec;62(6):1510-22.

QSM.tsvd — Function

tsvd(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    pad::NTuple{3, Integer} = (0, 0, 0),
    bdir::NTuple{3, Real} = (0, 0, 1),
    Dkernel::Symbol = :k,
    thr::Real = 0.15,
) -> typeof(similar(f))

Truncated singular value decomposition (TSVD) [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

thr::Real = 0.15: threshold for k-space filter

Returns

typeof(similar(f)): susceptibility map

References

[1] Wharton S, Schäfer A, Bowtell R. Susceptibility mapping in the human brain using threshold‐based k‐space division. Magnetic resonance in medicine. 2010 May;63(5):1292-304.

QSM.tv — Function

tv(
    f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)},
    mask::AbstractArray{Bool, 3},
    vsz::NTuple{3, Real};
    W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing,
    Wtv::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing,
    pad::NTuple{3, Integer} = (0, 0, 0),
    Dkernel::Symbol = :k,
    bdir::NTuple{3, Real} = (0, 0, 1),
    lambda::Real = 1e-3,
    rho::Real = 100*lambda,
    mu::Real = 1,
    tol::Real = 1e-3,
    maxit::Integer = 250,
    verbose::Bool = false,
) -> typeof(similar(f))

Total variation deconvolution using ADMM [1].

Arguments

f::AbstractArray{<:AbstractFloat, N ∈ (3, 4)}: unwrapped (multi-echo) local field/phase

mask::AbstractArray{Bool, 3}: binary mask of region of interest

vsz::NTuple{3, Real}: voxel size

Keywords

W::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, N)}} = nothing: data fidelity weights

Wtv::Union{Nothing, AbstractArray{<:AbstractFloat, M ∈ (3, 5)}} = nothing: total variation weights

M = 3: same weights for all three gradient directions and all echoes

M = 4 = N: same weights for all three gradient directions, different weights for echoes

M = 5, (size(Wtv)[4,5] = [1 or N, 3]: different weights for each gradient direction

pad::NTuple{3, Integer} = (0, 0, 0): zero padding array

< 0: no padding

≥ 0: minimum padding to fast fft size

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

Dkernel::Symbol = :k: dipole kernel method

lambda::Real = 1e-3: regularization parameter

rho::Real = 100*lambda: Lagrange multiplier penalty parameter

mu::Real = 1: Lagrange multiplier penalty parameter (unused if W = nothing)

tol::Real = 1e-3: stopping tolerance

maxit::Integer = 250: maximum number of iterations

verbose::Bool = false: print convergence information

Returns

typeof(similar(f)): susceptibility map

References

[1] Bilgic B, Fan AP, Polimeni JR, Cauley SF, Bianciardi M, Adalsteinsson E, Wald LL, Setsompop K. Fast quantitative susceptibility mapping with L1‐regularization and automatic parameter selection. Magnetic resonance in medicine. 2014 Nov;72(5):1444-59.

Binary masks

QSM.bet — Function

bet(
    mag::AbstractArray{<:Real, 3},
    vsz::NTuple{3, Real},
    betargs::AbstractString = "-m -n -f 0.5"
) -> Array{Bool, 3}

Interface to FSL's bet.

Arguments

mag::AbstractArray{<:Real, 3}: magnitude image

vsz::NTuple{3, Real}: voxel size

betargs::AbstractString = "-m -n -f 0.5": bet options

Returns

Array{Bool, 3}: binary brain mask

QSM.crop_indices — Function

crop_indices(x::AbstractArray, out::Number = 0) -> CartesianIndices

Indices to crop mask.

Arguments

x::AbstractArray: mask

out::Number = 0: value in x considered outside

Returns

CartesianIndices: indices to crop mask

QSM.crop_mask — Function

crop_mask(
    x::AbstractArray,
    m::AbstractArray = x;
    out::Number = 0
) -> typeof(x[...])

Crop array to mask.

Arguments

x::AbstractArray: array to be cropped

m::AbstractArray: mask

Keywords

out::Number = 0: value in m considered outside

Returns

typeof(x[...]): cropped array

QSM.erode_mask — Function

erode_mask(mask::AbstractArray{Bool, 3}, iter::Integer = 1) -> typeof(similar(mask))

Erode binary mask using an 18-stencil cube.

Arguments

mask::AbstractArray{Bool, 3}: binary mask

iter::Integer = 1: erode iter times

Returns

typeof(similar(mask)): eroded binary mask

QSM.erode_mask! — Function

erode_mask!(
    emask::AbstractArray{Bool, 3},
    mask::AbstractArray{Bool, 3},
    iter::Integer = 1
) -> emask

Erode binary mask using an 18-stencil cube.

Arguments

emask::AbstractArray{Bool, 3}: eroded binary mask

mask::AbstractArray{Bool, 3}: binary mask

iter::Integer = 1: erode iter times

Returns

emask: eroded binary mask

Kernels

QSM.dipole_kernel — Function

dipole_kernel(sz, vsz; kwargs...) =
    dipole_kernel(Float64, sz, vsz; kwargs...)

dipole_kernel(
    ::Type{T<:AbstractFloat},
    sz::NTuple{3, Integer},
    vsz::NTuple{3, Real};
    bdir::NTuple{3, Real} = (0, 0, 1),
    method::Symbol = :kspace,
    dsz::NTuple{3, Integer} = sz,
    transform::Union{Nothing, Symbol} = nothing,
    shift::Bool = false
) -> Array{T, 3}

Dipole kernel.

By default the dipole kernel is constructed in k-space and centered at index (1,1,1).

Arguments

sz::NTuple{3, Integer}: array size

vsz::NTuple{3, Real}: voxel size

Keywords

bdir::NTuple{3, Real} = (0, 0, 1): unit vector of B field direction

method::Symbol = :kspace: create in image space (:i, :ispace) or kspace (:k, :kspace)

dsz::NTuple{3, Integer} = sz:

method ∈ (:k, :kspace): unused

method ∈ (:i, :ispace): dipole kernel size in image space

transform::Union{Nothing, Symbol} = nothing:

method ∈ (:k, :kspace): create :rfft or :fft kspace dipole kernel

method ∈ (:i, :ispace): transform image space dipole kernel (:rfft, :fft)

shift::Bool = false:

method ∈ (:k, :kspace): kernel centered at (1,1,1) (false) or N÷2+1 (true)

method ∈ (:i, :ispace) and transform = nothing: kernel centered at N÷2+1 (false) or (1,1,1) (true)

method ∈ (:i, :ispace) and transform ∈ (:rfft, :fft): kernel centered at (1,1,1) (false) or N÷2+1 (true)

Returns

Array{T, 3}: dipole kernel

QSM.laplace_kernel — Function

laplace_kernel(vsz) =
    laplace_kernel(Float64, vsz)

laplace_kernel(sz, vsz; kwargs...) =
    laplace_kernel(Float64, sz, vsz; kwargs...)

laplace_kernel(::Type{T<:AbstractFloat}, vsz::NTuple{3, Real}) =
    laplace_kernel(T, (3, 3, 3), vsz, transform=nothing, shift=false)

laplace_kernel(
    ::Type{T<:AbstractFloat},
    sz::NTuple{3, Integer},
    vsz::NTuple{3, Real};
    negative::Bool = false,
    transform::Union{Nothing, Symbol} = nothing,
    shift::Bool = false
) -> Array{T, 3}

Discrete 7-point stencil Laplacian kernel.

Arguments

sz::NTuple{3, Integer}: array size

vsz::NTuple{3, Real}: voxel size

Keywords

negative::Bool = false: construct negative Laplacian (true)

transform::Union{Nothing, Symbol} = nothing: transform Laplacian kernel (:rfft, :fft)

shift::Bool = false:

transform = nothing: sphere centered at N÷2+1 (false) or (1,1,1) (true)

transform ∈ (:rfft, :fft): sphere centered at (1,1,1) (false) or N÷2+1 (true)

Returns

Array{T, 3}: Laplacian kernel

QSM.smv_kernel — Function

smv_kernel(sz, vsz, r; kwargs...) =
    smv_kernel(Float64, sz, vsz, r; kwargs...)

smv_kernel(
    ::Type{T<:AbstractFloat},
    sz::NTuple{3, Integer},
    vsz::NTuple{3, Real}
    r::Real;
    transform::Union{Nothing, Symbol} = nothing,
    shift::Bool = false
) -> Array{T, 3}

Spherical mean value kernel (SMV).

Arguments

sz::NTuple{3, Integer}: array size

vsz::NTuple{3, Real}: voxel size

r::Real: radius of sphere in units of vsz

Keywords

transform::Union{Nothing, Symbol} = nothing: transform SMV kernel (:rfft, :fft)

shift::Bool = false:

transform = nothing: sphere centered at N÷2+1 (false) or (1,1,1) (true)

transform ∈ (:rfft, :fft): sphere centered at (1,1,1) (false) or N÷2+1 (true)

Returns

Array{T, 3}: SMV kernel

Multi-echo

QSM.fit_echo_linear — Function

fit_echo_linear(
    phas::AbstractArray{<:AbstractFloat, N > 1},
    W::AbstractArray{<:AbstractFloat, N > 1},
    TEs::NTuple{NT > 1, Real};
    phase_offset::Bool = true
) -> Tuple{typeof(similar(phas)){N-1}, [typeof(similar(phas)){N-1}]}

Weighted least squares for multi-echo data.

Arguments

phas::AbstractArray{<:AbstractFloat, N > 1}: unwrapped multi-echo phase

W::AbstractArray{<:AbstractFloat, N > 1}: reciprocal of error variance of voxel

TEs::NTuple{NT > 1, Real}: echo times

Keywords

phase_offset::Bool = true: model phase offset (true)

Returns

typeof(similar(phas)){N-1}: weighted least-squares estimate for phase

[typeof(similar(phas)){N-1}]: weighted least-squares estimate for phase offset if phase_offset = true

QSM.fit_echo_linear! — Function

fit_echo_linear!(
    p::AbstractArray{<:AbstractFloat, N},
    phas::AbstractArray{<:AbstractFloat, M > 1},
    W::AbstractArray{<:AbstractFloat, M > 1},
    TEs::NTuple{NT > 1, Real}
) -> p

Weighted least squares for multi-echo data (phase offset = 0).

Arguments

p::AbstractArray{<:AbstractFloat, N}: weighted least-squares estimate for phase

phas::AbstractArray{<:AbstractFloat, M > 1}: unwrapped multi-echo phase

W::AbstractArray{<:AbstractFloat, M > 1}: reciprocal of error variance of voxel

TEs::NTuple{NT > 1, Real}: echo times

Returns

p: weighted least-squares estimate for phase

fit_echo_linear!(
    p::AbstractArray{<:AbstractFloat, N},
    p0::AbstractArray{<:AbstractFloat, N},
    phas::AbstractArray{<:AbstractFloat, M > 1},
    W::AbstractArray{<:AbstractFloat, M > 1},
    TEs::NTuple{NT > 1, Real}
) -> (p, p0)

Weighted least squares for multi-echo data (estimate phase offset).

Arguments

p::AbstractArray{<:AbstractFloat, N}: weighted least-squares estimate for phase

p0::AbstractArray{<:AbstractFloat, N}: weighted least-squares estimate for phase offset

phas::AbstractArray{<:AbstractFloat, M > 1}: unwrapped multi-echo phase

W::AbstractArray{<:AbstractFloat, M > 1}: reciprocal of error variance of voxel

TEs::NTuple{NT > 1, Real}: echo times

Returns

p: weighted least-squares estimate for phase

p0: weighted least-squares estimate for phase offset

Other

R2* mapping

QSM.r2star_arlo — Function

function r2star_arlo(
    mag::AbstractArray{<:AbstractFloat, N > 1},
    TEs::NTuple{NT > 1, Real},
    mask::Union{Nothing, AbstractArray{Bool}} = nothing
) -> typeof(similar(mag, size(mag)[1:N-1]))

Auto-Regression on Linear Operations (ARLO) [1].

Arguments

mag::AbstractArray{<:AbstractFloat, N > 1}: multi-echo magnitude

TEs::NTuple{NT > 1, Real}: echo times

mask::Union{Nothing, AbstractArray{Bool}} = nothing: binary mask of region of interest

Returns

typeof(similar(mag, size(mag)[1:N-1])): R2* map (1 / units of TEs)

References

[1] Pei M, Nguyen TD, Thimmappa ND, Salustri C, Dong F, Cooper MA, Li J, Prince MR, Wang Y. Algorithm for fast monoexponential fitting based on auto‐regression on linear operations (ARLO) of data. Magnetic resonance in medicine. 2015 Feb;73(2):843-50.

QSM.r2star_crsi — Function

function r2star_crsi(
    mag::AbstractArray{<:AbstractFloat, N > 1},
    TEs::NTuple{NT > 1, Real},
    mask::Union{Nothing, AbstractArray{Bool}} = nothing;
    M::Integer = 3,
    sigma::Union{Nothing, Real} = nothing,
    Rsz::NTuple{N-1, Integer} = size(mag)[1:N-1] .÷ 20,
) -> typeof(similar(mag, size(mag)[1:N-1]))

Calculation of Relaxivities by Signal Integration (CRSI) [1].

Arguments

mag::AbstractArray{<:AbstractFloat, N > 1}: multi-echo magnitude

TEs::NTuple{NT > 1, Real}: echo times

mask::Union{Nothing, AbstractArray{Bool}} = nothing: binary mask of region of interest

Keywords

M::Integer = 3: interpolation factor

sigma::Union{Nothing, Real} = nothing: noise

Rsz::NTuple{N-1, Integer} = size(mag)[1:N-1] .÷ 20:

sigma isa Real: unused

sigma isa Nothing: size of kernels used to calculate the noise from the background signal of the magnitude.

Returns

typeof(similar(mag, size(mag)[1:N-1])): R2* map (1 / units of TEs)

References

[1] Song R, Loeffler RB, Holtrop JL, McCarville MB, Hankins JS, Hillenbrand CM. Fast quantitative parameter maps without fitting: Integration yields accurate mono‐exponential signal decay rates. Magnetic resonance in medicine. 2018 Jun;79(6):2978-85.

QSM.r2star_ll — Function

function r2star_ll(
    mag::AbstractArray{<:AbstractFloat, N > 1},
    TEs::NTuple{NT > 1, Real},
    mask::Union{Nothing, AbstractArray{Bool}} = nothing
) -> typeof(similar(mag, size(mag)[1:N-1]))

Log-linear fit.

Arguments

mag::AbstractArray{<:AbstractFloat, N > 1}: multi-echo magnitude

TEs::NTuple{NT > 1, Real}: echo times

mask::Union{Nothing, AbstractArray{Bool}} = nothing: binary mask of region of interest

Returns

typeof(similar(mag, size(mag)[1:N-1])): R2* map (1 / units of TEs)

QSM.r2star_numart2s — Function

function r2star_numart2s(
    mag::AbstractArray{<:AbstractFloat, N > 1},
    TEs::NTuple{NT > 1, Real},
    mask::Union{Nothing, AbstractArray{Bool}} = nothing
) -> typeof(similar(mag, size(mag)[1:N-1]))

Numerical Algorithm for Real-time T2* mapping (NumART2*) [1].

Arguments

mag::AbstractArray{<:AbstractFloat, N > 1}: multi-echo magnitude

TEs::NTuple{NT > 1, Real}: echo times

mask::Union{Nothing, AbstractArray{Bool}} = nothing: binary mask of region of interest

Returns

typeof(similar(mag, size(mag)[1:N-1])): R2* map (1 / units of TEs)

References

[1] Hagberg GE, Indovina I, Sanes JN, Posse S. Real‐time quantification of T2* changes using multiecho planar imaging and numerical methods. Magnetic Resonance in Medicine: An Official Journal of the International Society for Magnetic Resonance in Medicine. 2002 Nov;48(5):877-82.

Finite differences

QSM.gradfp — Function

gradfp(
    u::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real}
) -> NTuple{3, typeof(similar(u))}

First order forward difference gradient with periodic boundaries.

Arguments

u::AbstractArray{<:AbstractFloat, 3}: input array

h::NTuple{3, Real}: grid spacing

Returns

dx::typeof(similar(u)): x-component of gradient of u

dy::typeof(similar(u)): y-component of gradient of u

dz::typeof(similar(u)): z-component of gradient of u

QSM.gradfp! — Function

gradfp!(
    dx::AbstractArray{<:AbstractFloat, 3},
    dy::AbstractArray{<:AbstractFloat, 3},
    dz::AbstractArray{<:AbstractFloat, 3},
    u::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real},
) -> (dx, dy, dz)

First order forward difference gradient with periodic boundaries.

Arguments

dx::AbstractArray{<:AbstractFloat, 3}: x-component of gradient of u

dy::AbstractArray{<:AbstractFloat, 3}: y-component of gradient of u

dz::AbstractArray{<:AbstractFloat, 3}: z-component of gradient of u

u::AbstractArray{<:AbstractFloat, 3}: input array

h::NTuple{3, Real}: grid spacing

Returns

dx: x-component of gradient of u

dy: y-component of gradient of u

dz: z-component of gradient of u

QSM.gradfp_adj — Function

gradfp_adj(
    dx::AbstractArray{<:AbstractFloat, 3},
    dy::AbstractArray{<:AbstractFloat, 3},
    dz::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real}
) -> typeof(similar(dx, promote_eltype(dx, dy, dz)))

Adjoint of first order forward difference gradient with periodic boundaries.

Arguments

dx::AbstractArray{<:AbstractFloat, 3}: x-component

dy::AbstractArray{<:AbstractFloat, 3}: y-component

dz::AbstractArray{<:AbstractFloat, 3}: z-component

h::NTuple{3, Real}: grid spacing

Returns

u::typeof(similar(dx, promote_eltype(dx, dy ,dz))): divergence of [dx, dy, dz]

QSM.gradfp_adj! — Function

gradfp_adj!(
    u::AbstractArray{<:AbstractFloat, 3}
    dx::AbstractArray{<:AbstractFloat, 3},
    dy::AbstractArray{<:AbstractFloat, 3},
    dz::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real}
) -> u

Adjoint of first order forward difference gradient with periodic boundaries.

Arguments

u::AbstractArray{<:AbstractFloat, 3}: divergence of [dx, dy, dz]

dx::AbstractArray{<:AbstractFloat, 3}: x-component

dy::AbstractArray{<:AbstractFloat, 3}: y-component

dz::AbstractArray{<:AbstractFloat, 3}: z-component

h::NTuple{3, Real}: grid spacing

Returns

u: divergence of [dx, dy, dz]

QSM.lap — Function

lap(
    u::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real}
) -> typeof(similar(u))

Second order central difference Laplacian.

Arguments

u::AbstractArray{<:AbstractFloat, 3}: input array

h::NTuple{3, Real}: grid spacing

Returns

d2u::typeof(similar(u)): discrete Laplacian of u

QSM.lap! — Function

lap!(
    d2u::AbstractArray{<:AbstractFloat, 3},
    u::AbstractArray{<:AbstractFloat, 3},
    h::NTuple{3, Real}
) -> d2u

Second order central difference Laplacian.

Arguments

d2u::AbstractArray{<:AbstractFloat, 3}: discrete Laplacian of u

u::AbstractArray{<:AbstractFloat, 3}: input array

h::NTuple{3, Real}: grid spacing

Returns

d2u: discrete Laplacian of u

Padding

QSM.fastfftsize — Function

fastfftsize(
    sz::NTuple{N, Integer},
    ksz::NTuple{M, Integer} = ntuple(_ -> 0, Val(N));
    rfft::Bool = false
) -> NTuple{N, Integer}

Next fast fft size greater than or equal to sz for convolution with a kernel of size ksz.

Arguments

x::AbstractArray{T, N}: array to pad

ksz::NTuple{M, Integer} = ntuple(_ -> 0, Val(N)): convolution kernel size

ksz[n] < 0: no padding for dimension n

Keywords

rfft::Bool = false: force first dimension to be even (true)

Returns

NTuple{N, Integer}: fast fft size

QSM.padarray! — Function

padarray!(
    xp::AbstractArray{Txp, N},
    x::AbstractArray{Tx, N},
    pad::Symbol = :fill,
    val::Number = 0
) -> xp

Pad array keeping it centered at n÷2+1.

Arguments

xp::AbstractArray{Txp, N}: padded array

x::AbstractArray{Tx, N}: array to pad

pad::Symbol = :fill: padding method

:fill

:circular

:replicate

:symmetric

:reflect

val::Number = 0: pads array with val if pad = :fill

Returns

xp: padded array

QSM.padfastfft — Function

padfastfft(
    x::AbstractArray{T, N},
    ksz::NTuple{M, Integer} = ntuple(_ -> 0, Val(N));
    pad::Symbol = :fill,
    val::Number = zero(T),
    rfft::Bool = false,
) -> typeof(similar(x, szp))

Pad array x to a fast fft size for convolution with a kernel of size ksz, keeping the array centered at n÷2+1.

Arguments

x::AbstractArray{T, N}: array to pad

ksz::NTuple{M, Integer} = ntuple(_ -> 0, Val(N)): convolution kernel size

ksz[n] < 0: no padding for dimension n

Keywords

pad::Symbol = :fill: padding method

:fill

:circular

:replicate

:symmetric

:reflect

val::Number = 0: pads array with val if pad = :fill

rfft::Bool = false: force first dimension to be even (true)

Returns

typeof(similar(x, szp)): padded array

QSM.unpadarray — Function

unpadarray(
    xp::AbstractArray{T, N},
    sz::NTuple{N, Integer}
) -> typeof(similar(xp, sz))

Extract array of size sz centered at n÷2+1 from xp.

QSM.unpadarray! — Function

unpadarray!(
    x::AbstractArray{Tx, N},
    xp::AbstractArray{Txp, N},
    sz::NTuple{N, Integer}
) -> x

Extract array centered at n÷2+1 from xp into x.

Miscellaneous

QSM.psf2otf — Function

psf2otf(
    psf::AbstractArray{<:Number, N},
    sz::NTuple{N, Integer} = size(psf);
    rfft::Bool = false,
) -> otf

Implementation of MATLAB's psf2otf function.

Arguments

psf::AbstractArray{T<:Number, N}: point-spread function

sz::NTuple{N, Integer}: size of output array; must not be smaller than psf

Keywords

rfft::Bool = false:

T<:Real: compute fft (false) or rfft (true)

T<:Complex: unused

Returns

otf: optical transfer function